Create preview url rule when deployment is created
